refactor: move statistics into extension
