+++ src/Reactant.jl
@@ -1,7 +1,6 @@
 module Reactant
 
 using PackageExtensionCompat
-using Statistics: Statistics
 
 include("mlir/MLIR.jl")
 include("XLA.jl")
